tela login e permissoes

+++ src/middlewares/users.ts
@@ -1,5 +1,6 @@
 import { connect } from '../database'
 import { RequestHandler } from "express"
+import jwt, { JwtPayload } from "jsonwebtoken"
 
 const getManyUsers: RequestHandler = async (req, res) => {
   const db = await connect()
@@ -10,26 +11,51 @@ const getManyUsers: RequestHandler = async (req, res) => {
 const createUser: RequestHandler = async (req, res) => {
   const db = await connect()
   const { name, email, password } = req.body
+
   const result = await db.run('INSERT INTO users (name, email, password) VALUES (?, ?, ?)', [name, email, password])
   const user = await db.get('SELECT id, name, email FROM users WHERE id = ?', [result.lastID])
+  
   res.json(user)
 }
 
 const updateUser: RequestHandler = async (req, res) => {
-  const db = await connect()
-  const { name, email } = req.body
-  const { id } = req.params
-  await db.run('UPDATE users SET name = ?, email = ? WHERE id = ?', [name, email, id])
-  const user = await db.get('SELECT * FROM users WHERE id = ?', [id])
-  res.json(user)
-}
+    const db = await connect();
+    const { name, email, password } = req.body;
+    const { id } = req.params;
+
+    if (!req.headers.authorization)
+        return res.status(401).json({ message: 'No token provided' })
+    
+    const token = req.headers.authorization
+    const decoded = jwt.decode(token) as JwtPayload;
+
+    if (id != decoded.id) {
+        return res.status(403).json({ message: 'Você não pode editar dados de outros coleguinhas usuarios.' });
+    }
+
+    await db.run('UPDATE users SET name = ?, email = ?, password = ? WHERE id = ?', [name, email, password, id]);
+    const user = await db.get('SELECT * FROM users WHERE id = ?', [id]);
+    res.json(user);
+};
+
 
 const deleteUser: RequestHandler = async (req, res) => {
-  const db = await connect()
-  const { id } = req.params
-  await db.run('DELETE FROM users WHERE id = ?', [id])
-  res.json({ message: 'User deleted' })
-}
+    const db = await connect();
+    const { id } = req.params;
+
+    if (!req.headers.authorization)
+        return res.status(401).json({ message: 'No token provided' })
+    
+    const token = req.headers.authorization
+    const decoded = jwt.decode(token) as JwtPayload;
+
+    if (id != decoded.id) {
+        return res.status(403).json({ message: 'Não pode excluir dados de outros coleguinhas usuarios.' });
+    }
+
+    await db.run('DELETE FROM users WHERE id = ?', [id]);
+    res.json({ message: 'Usuário excluído' });
+};
 
 export default { 
   getManyUsers,
